requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
scikit-learn
tqdm
requests
python-dotenv
